Completely refactor to decouple the agents, neural nets, and environments
ensime-sbt plugin added for use with ensime-vim, Add scalamock for
mocking and stubbing objects in unit tests, Add a couple of unit tests.
 The project builds and tests run, but tests are currently failing.
