+++ project/plugins.sbt
@@ -1,2 +1,3 @@
   addSbtPlugin("org.brianmckenna" % "sbt-wartremover" % "0.14")
   addSbtPlugin("com.sksamuel.scapegoat" %% "sbt-scapegoat" % "1.0.0")
+  addSbtPlugin("org.ensime" % "ensime-sbt" % "0.1.7")
